backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-generativeai==0.3.2
google-cloud-storage==2.14.0
firebase-admin==6.3.0
python-multipart==0.0.6
pydantic==2.5.3
